Move the files
